getBooksById move to BookService + findByIdorThrowException updated to handle both execptions for categories and books
