+++ src/main/java/com/onlinebookstore/onlinebookstore/service/impl/BookServiceImpl.java
@@ -1,5 +1,6 @@
 package com.onlinebookstore.onlinebookstore.service.impl;
 
+import com.onlinebookstore.onlinebookstore.dto.book.BookDtoWithoutCategoriesIds;
 import com.onlinebookstore.onlinebookstore.dto.book.BookRequestDto;
 import com.onlinebookstore.onlinebookstore.dto.book.BookResponseDto;
 import com.onlinebookstore.onlinebookstore.exeption.EntityNotFoundException;
@@ -46,13 +47,13 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public BookResponseDto findById(Long id) {
-        Book book = findByIdOrThrowException(id);
+        Book book = findByIdOrThrowException(id, Book.class);
         return bookMapper.toDto(book);
     }
 
     @Override
     public BookResponseDto update(Long id, BookRequestDto requestDto) {
-        Book book = findByIdOrThrowException(id);
+        Book book = findByIdOrThrowException(id, Book.class);
         bookMapper.updateFromDto(requestDto, book);
         setCategories(book, requestDto.getCategoryIds());
         Book updatedBook = bookRepository.save(book);
@@ -68,9 +69,32 @@ public class BookServiceImpl implements BookService {
         }
     }
 
-    private Book findByIdOrThrowException(Long id) {
-        return bookRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Can't find book by id " + id));
+    @Override
+    public List<BookDtoWithoutCategoriesIds> getBooksById(Long id, Pageable pageable) {
+        Category category = findByIdOrThrowException(id, Category.class);
+        List<BookDtoWithoutCategoriesIds> books = bookRepository
+                .findAllByCategoriesId(id, pageable).stream()
+                .map(bookMapper::toDtoWithoutCategories)
+                .collect(Collectors.toList());
+        if (books.isEmpty()) {
+            throw new EntityNotFoundException("Can't find any books for category id " + id);
+        }
+        return books;
+    }
+
+    private <T> T findByIdOrThrowException(Long id, Class<T> entityClass) {
+        if (entityClass.equals(Book.class)) {
+            return (T) bookRepository.findById(id)
+                    .orElseThrow(()
+                            -> new EntityNotFoundException("Can't find book by id " + id));
+        } else if (entityClass.equals(Category.class)) {
+            return (T) categoryRepository.findById(id)
+                    .orElseThrow(()
+                            -> new EntityNotFoundException("Can't find category by id " + id));
+        } else {
+            throw new IllegalArgumentException("Unsupported entity class: "
+                    + entityClass.getName());
+        }
     }
 
     private void setCategories(Book book, Set<Long> categoryIds) {
