+++ src/main/java/com/onlinebookstore/onlinebookstore/service/impl/CategoryServiceImpl.java
@@ -1,13 +1,10 @@
 package com.onlinebookstore.onlinebookstore.service.impl;
 
-import com.onlinebookstore.onlinebookstore.dto.book.BookDtoWithoutCategoriesIds;
 import com.onlinebookstore.onlinebookstore.dto.category.CategoryRequestDto;
 import com.onlinebookstore.onlinebookstore.dto.category.CategoryResponseDto;
 import com.onlinebookstore.onlinebookstore.exeption.EntityNotFoundException;
-import com.onlinebookstore.onlinebookstore.mapper.BookMapper;
 import com.onlinebookstore.onlinebookstore.mapper.CategoryMapper;
 import com.onlinebookstore.onlinebookstore.model.Category;
-import com.onlinebookstore.onlinebookstore.repository.BookRepository;
 import com.onlinebookstore.onlinebookstore.repository.CategoryRepository;
 import com.onlinebookstore.onlinebookstore.service.interfaces.CategoryService;
 import java.util.List;
@@ -20,9 +17,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
-    private final BookRepository bookRepository;
     private final CategoryMapper categoryMapper;
-    private final BookMapper bookMapper;
 
     @Override
     public CategoryResponseDto getById(Long id) {
@@ -35,19 +30,6 @@ public class CategoryServiceImpl implements CategoryService {
         return categoryRepository.findAll(pageable).stream()
                 .map(categoryMapper::toDto)
                 .collect(Collectors.toList());
-    }
-
-    @Override
-    public List<BookDtoWithoutCategoriesIds> getBooksById(Long id, Pageable pageable) {
-        Category category = findByIdOrThrowException(id);
-        List<BookDtoWithoutCategoriesIds> books = bookRepository
-                .findAllByCategoriesId(id, pageable).stream()
-                .map(bookMapper::toDtoWithoutCategories)
-                .collect(Collectors.toList());
-        if (books.isEmpty()) {
-            throw new EntityNotFoundException("Can't find any books for category id " + id);
-        }
-        return books;
     }
 
     @Override
